services/vision-plante/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
ultralytics>=8.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
kafka-python>=2.0.2
python-json-logger>=2.0.0
aiofiles>=23.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
minio>=7.1.0
shapely>=2.0.0
